Move functions to english_compare.py
